+++ test/widgets/event_all_day_test.dart
@@ -23,8 +23,8 @@ void main() {
 
   final interaction = CalendarInteraction(
     inputMode: InputMode.precise,
-    createEventGesture: CreateEventGesture.tap,
-    modifyEventGesture: CreateEventGesture.tap,
+    createEventGesture: EventInteractionGesture.tap,
+    modifyEventGesture: EventInteractionGesture.tap,
   );
 
   setUp(() {