+++ test/widgets/multi_day_rule_on_view_test.dart
@@ -17,8 +17,8 @@ void main() {
 
   final interaction = CalendarInteraction(
     inputMode: InputMode.precise,
-    createEventGesture: CreateEventGesture.tap,
-    modifyEventGesture: CreateEventGesture.tap,
+    createEventGesture: EventInteractionGesture.tap,
+    modifyEventGesture: EventInteractionGesture.tap,
   );
 
   setUp(() {